src/tablaAsignacion.py
class tablaAsignacion:
    
    def __init__(self):
        self.tabla = [  
            "T",
            "R",
            "W",
            "A",
            "G",
            "M",
            "Y",
            "F",
            "P",
            "D",
            "X",
            "B",
            "N",
            "J",
            "Z",
            "S",
            "Q",
            "V",
            "H",
            "L",
            "C",
            "K",
            "E",
        ]
        
    def getTabla(self):
        return self.tabla

    def getLetra(self,posicion):
        return self.tabla[posicion]
    
    def getLength(self):
        return len(self.tabla)